requirements.txt
Django>=3.0.10,<3.1.0
django-filter==2.3.0
djangorestframework==3.11.1
psycopg2-binary==2.8.6
django-celery-beat==2.0.0
Celery==4.4.7
feedparser==6.0.1
redis==3.5.3
coreapi==2.3.3
pyyaml==5.3.1
coverage==5.3